refactor: introduce common crate
