+++ hello-tracing-backend/src/main.rs
@@ -1,5 +1,4 @@
 mod api;
-mod otel;
 
 use anyhow::{Context, Result};
 use configured::Configured;
