+++ hello-tracing-common/src/otel/mod.rs
@@ -1,0 +1,2 @@
+pub mod grpc;
+pub mod http;
